backend/requirements.txt
# JanShakti.AI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-multipart==0.0.12
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
nltk==3.9.1

# Sentiment Analysis
transformers==4.45.0
torch==2.4.1

# YOLOv8 Computer Vision
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# Speech-to-Text
openai-whisper==20240930

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
